Refactor and update requirements
