--- app/requirements.txt
@@ -1,89 +1,0 @@
-APScheduler==3.5.1
-asn1crypto==0.24.0
-bcrypt==3.1.4
-beautifulsoup4==4.6.0
-cachetools==2.1.0
-certifi==2018.1.18
-cffi==1.11.5
-chardet==3.0.4
-click==6.7
-cryptography==2.2.2
-dill==0.2.7.1
-docopt==0.6.2
-enum34==1.1.6
-Flask==0.12.2
-Flask-HTTPAuth==3.2.3
-Flask-SQLAlchemy==2.3.2
-funcsigs==1.0.2
-future==0.16.0
-futures==3.2.0
-gapic-google-cloud-datastore-v1==0.15.3
-gapic-google-cloud-error-reporting-v1beta1==0.15.3
-gapic-google-cloud-logging-v2==0.91.3
-google-api-core==0.1.4
-google-api-python-client==1.6.6
-google-auth==1.4.1
-google-cloud==0.32.0
-google-cloud-bigquery==0.28.0
-google-cloud-bigquery-datatransfer==0.1.1
-google-cloud-bigtable==0.28.1
-google-cloud-container==0.1.1
-google-cloud-core==0.28.1
-google-cloud-datastore==1.4.0
-google-cloud-dns==0.28.0
-google-cloud-error-reporting==0.28.0
-google-cloud-firestore==0.28.0
-google-cloud-language==1.0.2
-google-cloud-logging==1.4.0
-google-cloud-monitoring==0.28.1
-google-cloud-pubsub==0.30.1
-google-cloud-resource-manager==0.28.1
-google-cloud-runtimeconfig==0.28.1
-google-cloud-spanner==0.29.0
-google-cloud-speech==0.30.0
-google-cloud-storage==1.6.0
-google-cloud-trace==0.17.0
-google-cloud-translate==1.3.1
-google-cloud-videointelligence==1.0.1
-google-cloud-vision==0.29.0
-google-gax==0.15.16
-google-resumable-media==0.3.1
-googleapis-common-protos==1.5.3
-grpc-google-iam-v1==0.11.4
-grpcio==1.12.0
-httplib2==0.11.3
-idna==2.6
-ipaddress==1.0.19
-itsdangerous==0.24
-Jinja2==2.10
-MarkupSafe==1.0
-numpy==1.14.2
-oauth2client==3.0.0
-pandas==0.23.0
-paramiko==2.4.1
-passlib==1.7.1
-ply==3.8
-proto-google-cloud-datastore-v1==0.90.4
-proto-google-cloud-error-reporting-v1beta1==0.15.3
-proto-google-cloud-logging-v2==0.91.3
-protobuf==3.5.2.post1
-psutil==5.4.5
-psycopg2==2.7.4
-pyasn1==0.4.2
-pyasn1-modules==0.2.1
-pycparser==2.18
-Pydap==3.2.2
-PyNaCl==1.2.1
-python-dateutil==2.7.3
-pytz==2018.3
-requests==2.18.4
-rsa==3.4.2
-simplekml==1.3.0
-singledispatch==3.4.0.3
-six==1.11.0
-SQLAlchemy==1.2.6
-tzlocal==1.5.1
-uritemplate==3.0.0
-urllib3==1.22
-WebOb==1.8.0
-Werkzeug==0.14.1
